pyproject.toml
[project]
name = "browser-use"
description = "Browser automation and interaction library"
authors = [
    { name = "Gregor Zunic" }
]
version = "0.1.26"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "MainContentExtractor>=0.0.4",
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.2",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "playwright>=1.49.0",
    "setuptools>=75.8.0",
    "pytest>=8.3.4",
    "posthog>=3.8.4",
    "pillow>=11.1.0",
    "langchain>=0.3.15",
    "langgraph>=0.2.66",
    "langchain-openai>=0.3.1",
    "pymupdf>=1.25.2",
]

[project.optional-dependencies]
dev = [
    "hatch>=1.13.0",
    "build>=1.2.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0"
]

[tool.ruff]
line-length = 130
select = ["E", "F", "I"]
fix = true

[tool.ruff.format]
quote-style = 'single'
indent-style = 'tab'
docstring-code-format = true

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
